requirements/base__dep.txt
pytz==2018.5  # https://github.com/stub42/pytz
python-slugify==1.2.5  # https://github.com/un33k/python-slugify
Pillow==5.2.0  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==18.1.0  # https://github.com/hynek/argon2_cffi
# redis>=2.10.6  # https://github.com/antirez/redis
redis==3.3.11  # https://github.com/antirez/redis
celery==4.2.1  # pyup: <5.0  # https://github.com/celery/celery
flower==0.9.3  # https://github.com/mher/flower
django-test-plus==1.0.22
python-dateutil
selenium==3.141.0
urllib3==1.25.7
cryptography>=3.0,<4

# Django
# ------------------------------------------------------------------------------
django==2.0.7  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.36.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.2  # https://github.com/django-compressor/django-compressor
# django-redis==4.9.0  # https://github.com/niwinz/django-redis
django-redis==4.10.0  # https://github.com/niwinz/django-redis
django-invitations==1.9.2
django-rest-invitations
django-phonenumber-field==1.3.0
django-anymail==2.2
django-guardian==2.2.0
djangorestframework-guardian
django-extensions==2.1.0  # https://github.com/django-extensions/django-extensions
django-import-export
django-simple-history

# Django REST Framework
djangorestframework==3.8.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
django-rest-auth==0.9.3
django-cors-headers==2.3.0
django-silk==3.0.0
django-celery-beat==2.0.0
djangorestframework-bulk

# Google Requirements
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-analytics-data
grpcio

fabric
boto3==1.13.25
aiohttp
asyncio
bs4